Add references.yml and autogen references for src/md files
This commit adds a YML bibliography references.yml for all
papers, standards, or implementations cited throughout the code
base.

Citations are normalized to the style `[@ID, ...]` in the source
code and to `[^ID]` in markdown files.

`autogen` is extended to scan the source and markdown files for
those annotations, and automatically adds:
- for source files, a header with all references made in the file,
- for markdown files, footnotes for all references

`autogen` also checks that each bibliography entry is used
at least once.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>