autogen: Generate BIBLIOGRAPHY.md with all citations
This commit extends `autogen` to emit a `BIBLIOGRAPHY.md`
rendering of `references.yml`. For each citation id, it also prints
a list of files that use this citation.

While in `autogen` we do also track the line numbers of the citations,
those are currently not printed in `BIBLIOGRAPHY.md` since it would
require re-generating `BIBLIOGRAPHY.md` too often.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>