CI: Run nix cache rebuild before running extended tests
This commit moves the nix workflow nix.yml into all.yml
and triggers the extended tests ci.yml only once the nix
workflow completes. This prevents concurrent rebuilding
of the expensive ci-cross shell. With the current patch,
the ci-cross shell is built and cached once in nix.yml,
and can then be reused by ci.yml.

Also, the dependency of the nix development tests on the
nix cache build for ci-cross is unnecessary and removed.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>