Add output bound for mulcache computation
Previously, we did not include a bounds assertion on the output of
`mlk_poly_mulcache_compute_native` since the AVX2 backend implemented
this as a no-op (it did not even initialize the mulcache).

Now that the AVX2 backend does use the mulcache in the same way as
the C and AVX2 code, we can add the output bound assertion.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>