FIPS202/AArch64: Add symbolic assembly for v8-A scalar Keccak ASM
A major challenge in developing fast assembly for Keccka-F1600 is
register pressure: The Keccak-F1600 state alone occupies 25 registers,
leaving only 5 for the remaining computation.

The AArch64 scalar assembly for Keccak-F1600 used in mlkem-native
is automatically derived from _symbolic_ assembly using the SLOTHY
superoptimizer: What was written by hand is a symbolic version of
the assembly using symbolic placeholders for registers, while the
concrete allocation was left to SLOTHY. Previously, however, only
the final assembly was part of the mlkem-native repository, lacking
both the symbolic versions and the instructions to run SLOTHY.
This negatively impacts transparency, reproducibility, and maintainability
of the library.

This commit adds the original symbolic AArch64 assembly for the
scalar implementation of Keccak-F1600, as well as a Makefile the
shows how to invoke SLOTHY to conduct the register allocation.
This Makefile is invoked as part of `autogen --slothy`, which in
turn is already part of the mlkem-native CI. Thus, with this commit,
the symbolic->concrete development flow is fully exercised in CI,
and the functional correctness of the result tested.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>