CI: Force testing of SLOTHY-optimized Keccak assembly
Prior to this commit, the SLOTHY CI would run `tests all --opt=OPT`
to test the freshly SLOTHY optimized assembly files.

With the introduction of Keccak into the scope of SLOTHY, this is
not sufficient, since the default choice of Keccak variant depends
on the target architecture. In particular, the x1-scalar Keccak would
not be tested by `tests all` run on Graviton4.

This commit refines the SLOTHY CI by forcing the x1 scalar Keccak
variant in the test. As further variants of Keccak are added to
the scope of SLOTHY, multiple invocations of `tests all` will be
needed.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>