FIPS202/AArch64: Add de-interleaved scalar/v8-A-Neon x4-fold Keccak
The x4-batched hybrid scalar/Neon implementation of Keccak on AArch64
is obtained by interleaving independent scalar and Neon implementations.
This interleaving was not done manually, but by SLOTHY. To this point,
however, the original non-interleaved source was not part of mlkem-native,
nor were the instructions for how to run SLOTHY to conduct the interleaving.
This impedes transparency, reproducibility, and maintenance.

This commit adds the de-interleaved version of the scalar/Neon Keccak
assembly for AArch64, as well as build instructions for how to run
SLOTHY to conduct the interleaving.

The CI is extended to test functional correctness of the freshly
interleaved code.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>