FIPS202/AArch64: Add de-interleaved scalar/v8-A/v8.4-A Keccak
Analogous to the previous commit:

The x4-batched hybrid scalar/Neon/Neon-v8.4-A implementation of Keccak
on AArch64 is obtained by interleaving independent scalar, Neon-v8-A
and Neon-v8.4-A(+SHA3) implementations. This interleaving was not done
manually, but by SLOTHY. To this point, however, the original non-interleaved
source was not part of mlkem-native, nor were the instructions for how
to run SLOTHY to conduct the interleaving. This impedes transparency,
reproducibility, and maintenance.

This commit adds the de-interleaved version of the scalar/Neon Keccak
assembly for AArch64, as well as build instructions for how to run
SLOTHY to conduct the interleaving.

The CI is extended to test functional correctness of the freshly
interleaved code.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>