HOL-Light: Autogenerate rejection sampling constant table
The HOL-Light proof script for AArch64 rejection sampling has
embedded in it the large precomputed constant table used by
the assembly. The same table is also embedded in the C source
of mlkem-native, and autogenerated from `scripts/autogen`.

To make it as obvious as possible that the constant table
assumed by the proof script matches the constant table used
by the source code, this commit extends `scripts/autogen`
to autogenerate the constant table embedded in the proof script
from the same Python code that also generates the constant table
in the C source.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>