Fix relicensing gap
Some files were erroneously and unintentionally left out of the
relicensing from Apache-2.0 to Apache-2.0 OR ISC OR MIT. This
commit fixes that.

See RELICENSE.md for the relicensing consent.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>

Relicensing META.sh, which I am the sole contributor to,
from the erroneously chosen CC-BY-4.0 to the
standard license `Apache-2.0 OR ISC OR MIT`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>