Remove output bound for poly_mulcache_compute
The higher level functions do not require a output bound for
poly_mulcache_compute.
However, all our backend (C/AVX2/Arm64) produce outputs in (-q,q) and hence
the lower level functions did have according assertions or post-conditions.

The HOL-Light proof for the AArch64 implementation of poly_mulcache_compute,
however, does not prove a output bound for it.
We are considering adding a post-condition, but that is some work:
https://github.com/pq-code-package/mlkem-native/issues/1032

For the meantime, this commit removes the post-condition from the lower level
functions such that it is consistent with the HOL-Light proof.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>
the contracts of the lower level functions