CBMC: Don't declare arrays in harnesses
When CBMC checks a function contract, it seems that the preconditions
set out in the contract are overlayed by the context set out in the
harness. For example, if a function takes a pointer to `foo`, and the
harness passes `&f` for `foo f`, then the function would never be
evaluated for the possibility of an invalid pointer being passed.
For that reason, harnesses must be as minimal as possible, merely
declaring uninitialized variables of the expected type and calling
the function under proof.

This commit fixes a few instances in mlkem-native where this was
not yet the case.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>