ACVP: Adjust to work with prompt + expectedResults rather than internal Projection
Thus far, we have been using the "internalProjection" files from
https://github.com/usnistgov/ACVP-Server/tree/master/gen-val/json-files.
These contain both the inputs and the expected outputs for the function under
test.

In real ACVP validation, the internalProjection is not available.
Rather, one gets a prompt containing the inputs and has to produce a
result json that has the match the expected results hold by the ACVP server.

This commit changes the ACVP client to consume two files separately:
the prompt file containing the inputs, and the expectedResults file containing
the outputs. Both files are available at the same location as the
internalProjection files.
At the end the results json is compared to the expectedResults json.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>