ACVP: Add --prompt, --output, --expected flags to acvp_client.py
This extends the ACVP client to take optional inputs --prompt, --output,
--expected.
If none of these inputs are provided, behaviour stays unchanged and the
testvectors in test/acvp_data are tested.

If a ACVP prompt json file is provided to the --prompt input, that prompt is
used instead.
If that is the case, the user has two options:
 1) Provide the expected results as a file path to --expected
 2) Provide a file path where the results should be written to without
   checking them.

(1) is for testing purposes; (2) is for actual ACVP certification purposes.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>