autogen: Automatically prefix assembly labels with function name
This commit extends `autogen` to go through labels in assembly files
and prefix them with the name of the function that is defined in the
file. The script can also detect and fixup minor reordering mistakes
in the prefix. Still, some manual patching had to be some for the
AVX2 assembly.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>