Examples: Minimize file import in `monolithic_build` example
The `monolithic_build` example does not need, and its build should
not rely on, any native files. To make sure this is tested, only
symlink non-native files into the example directory.

This breaks the example because of an issue in the SCU,
which was previously undetected. This will be fixed in the
following commit.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>