ACVP: Adjust to additionally support ACVTS data format
The ACVTS data structure is very slightly different from the sample files
in the usnistgov/ACVP-Server Github repository:
The prompt consists of a 2-element list, where the first element is
solely consisting of {"acvVersion": "1.0"} and the second element is
the usual prompt containing the test values.

This commit extends the ACVP client to automatically detect that case and
handle it properly.
The result is adjusted accordingly (getting the acvVersion appended).

I also disabled the alphabetic sorting of json keys as that just leads to
confusion. We are already using the same order as the NIST testvectors,
so there is no need to sort alphabetically.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>