Nix: Remove zig 0.10 and zig 0.11 shells and CI
zig0.11 seems to have dropped from the nix binary cache for MacOS.
This means it has to be built from source in the CI significanlty slowing down
CI.
Note that both zig 0.10 and zig 0.11 have been dropped in nixpkgs 25.05, so it
is expected that it is eventually dropped from the cache.
This commit removes zig 0.11 and the even older zig 0.10.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>