autogen: Clarify origin of `autogen` script
The `autogen` script emits a header indicating that the file under
construction is auto-generated.

However, consumers of mlkem-native are unlikely to copy the entire
mlkem-native repository, but rather only the main source tree. In
this case, the reference to `scripts/autogen` is confusing, because
that directory will likely not exist in the target project.

Improve this by stating thtat `scripts/autogen` refers to a path in
the mlkem-native repository.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>