Allow replacement of memcpy and memset
This commit extends the config to allow customization of
mlkem-native's dependencies on the stdlib: memset and memcpy.
It follows the same mechanism used for zeroize and randombytes.

Custom test configs are added illustrating how the new options work,
and they are exercised in CI.

Further, STDLIB.md is added, documenting the dependencies on the stdlib
and pointing to the customization mechanism.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>