ACVP: Download testvectors transparently on-the-fly
This commit changes the acvp_client.py to automatically download the required
testvectors file for a given version of the ACVP Server.
The main benefit of this change is that changes to the testvectors (which have
been happening much more frequent than I expected) do not pollute our git
history to the same extend anymore.
The local copies of the ACVP files are removed.

Note that before we were testing v1.0.0.40, v1.0.0.39, and v1.0.0.38.
This commit changes that to only use v1.0.0.40 by default.
Additional versions can be tested by calling the acvp_client.py directly
with the --version flag. This will be added to CI in the next commit.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>