Update CBMC proof guide and README to include installation instructions
We have seen numerous people trying to reproduce CBMC proofs with tool versions
not installed via nix. The results are mixed, but mostly a disaster.

This commit clarifies this in the README adding explicit installation
instructions and stressing that you _cannot_ expect to use any version and
proofs to work just fine.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>