Makefile: Set pipefail shellflag to increase robustness
As was reported in #386, `make run_kat` would previously succeed
even the `sha256sum` tool was not available on the host system.
The reason was that later commands in the pipe may shadow errors
in earlier commands (suchas 'command not found' errors in the case
of a non-existing sha256sum).

This commit fixes this by setting the `pipefail` shellflag
and forcing /bin/bash. META.sh already assumes /bin/bash, so
this is no restriction.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>