AArch64: Add ABI checker
This commit adds an ABI checker for all AArch64 assembly in mlkem-native.

The purpose of an ABI checker is to ensure that assembly files obey
the required calling conventions. For AArch64 specifically, it is
required by the AAPCS that x19-x30 and d8-d15 are unmodified.

It is important to test this since failure to obey the AAPCS can lead
to rare but fatal bugs depending on the exact compiler and compilation
settings.

On a high level, the ABI checker works as follows:

1. An assembly callstub

```c
  void asm_call_stub(struct register_state *input,
                     struct register_state *output,
                     void (*function_ptr)(void))
```

is implemented which calls the function specified by `function_ptr`
on the input register state `input`, and returns the output register
state in `output`. Importantly, it does _not_ assume that `function_ptr`
obeys the AAPCS, so must be implemented in assembly. This is done for
AArch64 only so far.

2. Randomized ABI checker

Based on `asm_call_stub`, we implement an ABI checker which calls
the target function with random register inputs and checks that the
output register state has retained the callee saved registers.

3. Pointer inputs

Most functions read/write memory specified via parameters. Calling
those functions with random input would lead to segmentation faults.
The ABI checker therefore does not randomize those arguments, but
instead sets up valid pointers of the right size for them.

For now, we always align the input pointers via MLK_ALIGN. A future
extension should take alignment constraints into account and vary
the alignment of the inputs within those constraints.

The ABI checker gets the information about which inputs are pointers
from the recently introduced YAML specification of AArch64 assembly.

Some implementation details:

- `autogen` is extended to generate one `test/abicheck/check_FUNCNAME.c`
  file per function under test. All tests are declared and enumerated
  in the autogenerated `test/abicheck/checks_all.h`.
- The handwritten `test/abicheck/abicheck.c` invokes all per-function
  ABI checks.
- We integrate the abicheck build by extending `test/mk/components.mk`
  with specific build rules. This allows us to partly piggy back on
  existing Makefile infrastructure, such as setting execution wrappers
  or cross prefixes. On the other hand, a lot of custom hackery is
  currently needed to get past the preprocessor directives guarding
  individual assembly files. This should be improved in the future.
- `scripts/tests` is extended to allow for `tests abicheck`, and
  `tests all` supports `--abicheck` and `--no-abicheck`. By default,
  the abicheck _is_ run with `tests all`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>