Makefile: Add compile-time compiler+host platform feature detection
Fixes #1144

This commit extends the Makefiles used for tests to check host compiler
and host CPU capabilities when setting archflags with AUTO=1.

This leads to SHA3 being enabled for valgrind CT tests in CI, which
breaks the test since valgrind does not support those instructions.
This is worked around by allowing the user to force values of the
compiler feature detection on the command line.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>