Runtime dispatch based on custom CPU capabilities function
This commit gives the user the ability to enable/disable the
use of native backends at runtime. Specifically, it introduces
a function mlk_is_native_capable() which backends call to check
if the host CPU provides the necessary features to run the native
backend. If not, the frontend falls back to the C version.

mlkem-native does not try to detect system capabilities at runtime
and hardcodes the capability function to 0, signaling that the backend
should be used, relying on feature detection / appropriate CFLAGS at
compile-time. Consumers wishing to build a single mlkem-native library
suitable for a variety of host system with heterogeneous capabilities,
MUST overwrite the runtime capability function accordingly.

Signed-off-by: dkostic <dkostic@amazon.com>
Signed-off-by: Rod Chapman <rodchap@amazon.com>
Signed-off-by: Hanno Becker <beckphan@amazon.com>