Mark sys.h as k-independent
In a previous version, this header shared macros with
mlkem_native.h and thus had to be declared as 'k-specific'
to be re-included in monolithic multilevel builds. This sharing
of macros seems to have been removed in the meantime, making it
possible to mark `sys.h` as k-indepednent, as it ought to be.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>