Integration/oqs: Autogenerate file list, pt 1
The script integration/liboqs/liboqs-check-filelist.py checks whether
the file lists in the libOQS YML files are up to date. However, until
now they still had to be edited manually when files were removed or
added.

This commit is the first of two removing said script and instead
auto-updating the file list in the libOQS YML files as part of
scripts/autogen. The check that the lists are up to date is then
part of autogen --dry-run.

This commit only re-formats the YAML files but does not yet update
the file list. This is to ease reviewing of the content changes.
The next commit will actually update the file list.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>