Introduce system capability enum and refine backend capability check
Previously, all backends would call the newly introduced
mlk_is_native_capable() function to check whether they should
be run or not.

This check is too coarse because different backends have different
requirements to check for. For example, the x86_64 backend requires
AVX2, while the AArch64 arithmetic backend should work on any Armv8-A
system, since Neon is mandatory for Armv8-A. On the other hand, the
FIPS202 AArch64 backend partially relies on the SHA3 extension of
Armv8.4-A.

This commit renames mlk_is_native_capable to mlk_sys_check_capability
and adds a parameter indicating the capability to check for. This is
drawn from the values of a newly introduced enum `mlk_sys_cap`.
The backends are rewritten to make use of this function.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>