Add further custom capability tests
This commit adds two custom configs suitable for Linux/AArch64 and
Linux/x86_64 systems, respectively. On Linux/AArch64, the config directly
reads from the ID_AA64PFR1_EL1 register to infer if the SHA3 extension is
supported. On Linux/x86_64, the CPUID instruction is used.

Those configurations are then exercised in the CI EC2 job, running on
c6g, c7g instances for AArch64, and t3 and t3a instances for x86_64.
Importantly, we force armv8.4-a+sha3 on c6g as well, and hence rely on
the runtime check to note that SHA3 is not actually supported, and hence
fall back to the C implementation.

To maximize sharing of CI logic, the running of custom configurations
is hoisted out into a separate github action which can then be invoked
from both ci.yml and ci_ec2_reusable.yml.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>