Enable constant-time declassifications in liboqs
This commit enables declassifications when constant-time tests are performed
in liboqs.
It is essential that MLK_CONFIG_CT_TESTING_ENABLED is set BEFORE sys.h
is included.
With this changed, the constant-time tests pass again in liboqs.

The reason that the constant time tests are not failing upstream is that
liboqs maintains a whitelist of code locations that are expected to fail
constant time testing.
In the meantime, we have ranamed checks_sk in
a00cc0de802ae1a45da312d398bd809d26378b65 which is causing constant-time
failures as the whitelist was not updated.
The whitelist approach is not robust and we should not rely on it.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>