HOL-Light: Add target to update byte code in proof scripts
This commit adds the `dump_bytecode` target to
`proofs/hol_light/arm/Makefile`, dumping the bytecode
of all AArch64 object files.

The output of this target is then used by the new command

```
autogen --update-hol-light-bytecode
```

to update the bytecode embedded in the various HOL-Light
proof scripts.

Note that this is purely convenience, not a soundness issue:
The proof scripts fail at runtime if the bytecode is not up to date.
Yet, any change to the assembly previously necessitated a manual
extraction of the bytecode in the format expected by HOL-Light,
which is cumbersome. Now, one only needs to run the above `autogen`
command once to update the byte code in the scripts.

The HOL-Light CI is updated to run

```
autogen --update-hol-light-bytecode --dry-run
```

before attempting any proof. Again, this is logically redundant
because each proof also checks the byte code, byte a) this offers
a fast-fail path avoiding running HOL-Light proofs when it's clear
that they will fail, b) it exercises `--update-hol-light-bytecode`
which is otherwise unused in CI.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>