Add invNTT backend unit test for constant input data
Implementation of the inverse NTT are at risk of overflow when not
correctly reducing those coefficients which grow exponentially fast
in the number of layers. Most notably, the first and second
coefficients in the output of the invNTT are sums of 128 input
coefficients each, which can easily overflow.

This commit tests this by running the invNTT on constant input data.
Note that since the invNTT often performs a reduction or normalization
at the beginning, it's unclear which exact choice of constant data
would trigger the largest coefficient growth, so we just try them all.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>