OQS CI: Remove upstream constant-time suppressions
In OQS constant-time testing is done by documenting all known potential
constant-time violations that are known to be false positives as
"passes".
In mlkem-native we take a diffeernt approach and have explicit declassifications
in the code.
The previous commit enabled the declassifications in the OQS CI.
However, as the "passes" are still in OQS upstream, tests would pass even if
our declassications would be removed.
This commit patches OQS as a part of our CI to remove the constant-time
passes.

I have tested in CI that this removal will make the constant-time tests
fail in case declassifcations are removed.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>