HOL-Light: Avoid hardcoded lengths in HOL-Light proof scripts
Previously, the HOL-Light proof scripts contained various hardcoded
constants related to the length of the code under proof, as well as
the lengths of preamble and postamble (stack spill and restore).

This commit replaces those hardcoded numerals with symbolic definitions.
As a result, SLOTHY re-optimization should now no longer require a proof
script change beyond the update of the byte code which is done by `autogen`
already.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>