HOL-Light: Run proofs until target PC is reached
Previously, the HOL-Light scripts hardcoded the exact number of
instructions executed, which was cumbersome to update by hand,
e.g. after re-running SLOTHY.

This commit adds a helper tactic `MAP_UNTIL_TARGET_PC` which
repeats a given tactic until the target PC is reached. In more
detail, it finds the `read PC s' = target_pc` in the postcondition
of the goal statement, and then repeats the tactic until there is
an assumption of the same shape.

Since the check is run with every step, and we often have hundreds
of assumptions, care is taken to immediately abort the search for
a suitable assumption once _some_ assumption of the above form is
found, regardless of whether the PC on the RHS match. For that
specific assumption, it is then checked whether RHS's match.
A plain `filter is_matching_pc_assumption ...`, in turn, would always
traverse all assumptions.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>