HOL-Light: Simplify proof of poly_tobytes
The precondition in the HOL-Light proof for mlkem_poly_tobytes
abstracts the memory contents at the source pointer as a list
of 16-bit words. However, it is only in the post-condition that
the length of this list is constrained to be 256.
This does not seem to increase the expressivity of the spec
but complicates the proof; in particular, it adds the need for
another tactic currently fed the total number of instructions
in the function, and which is not readily amenable to wrapping
by the recently introduced MAP_UNTIL_TARGET_PC function.

This commit moves the length constraints on the abstract list
of 16-bit values to the precondition, simplifying the proof
and thereby making it agnostic to the total number of instructions.

Note that in general there can be value to pushing preconditions
into premises of the post-condition: For example, in the NTT or
invNTT, one can shove the constraint that the provided twiddle
table should contain the right twiddles into a premise of the
post-condition. This strictly increases expressivity, because even
for wrong twiddle tables, one still gets safety properties from
the proof. However, in the case of poly_tobytes, there does not
seem to be such gain from moving the length constraint into the
post-condition.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>