Remove single occurrence of memcmp
The library uses an ordinary `memcmp` only once; in all other places, we
use our custom constant-time `mlk_ct_memcmp`. The one remaining occasion
of `memcmp` is not performance critical. To minimize dependency on the
standard library, this commit removes this single use of `memcmp` and
replaces it by `mlk_ct_memcmp`.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>