RV64: Allow arbitrary vector length, fallback to C if unsupported
This commit makes minor changes to the RV64 backend to make the
non-NTT component VLEN-agnostic. This mostly means using
__riscv_vsetvl_e16m1 instead of a hardcoded VLEN, plus some added
tail handling in rejection sampling.

For NTT and invNTT, the implementation is specific to VLEN=256,
and fallback to the C implementation for other vector lengths.
This should be improved in the future.

The compile-time guard for the RV64 backend is changed to checking
for RVV support only, but no specific VLEN. Similarly, the default
build flags use 128 rather then 256 as the minimum VLEN.

The README is adjusted accordingly. We also remove the warning
regarding the code being experimental, as it has by now undergone
thorough review and test.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>