Add -Wconversion to our builds
This commit enables further compiler warnings and adjusts the code
triggering the warnings. In some cases, we can work around the warnings
by tightening input/output types in function signatures. In other cases,
we explain in a comment why the truncation does not alter the value.
Finally, in some cases, the truncation can actually alter the value,
and we rely on signed->unsigned and unsigned->signed integer conversions.

Note that CBMC already checks that integer conversion does not alter
values, so the comments are a helpful indicator that we thought about
this, but not critical.

Resolves https://github.com/pq-code-package/mlkem-native/issues/1229

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>
Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>