Add MLK_CONFIG_SERIAL_FIPS202_ONLY option
Currently rejection sampling is performed using 4-way batched Keccak
regardless whether that is beneficial or not.
That approach requires to keep around 4 Keccak states in memory.
This approach is incompatible with using a Keccak accelerator where there is
only a single state and it lives inside of the accelerator.

This commit adds an option MLK_CONFIG_SERIAL_FIPS202_ONLY that switches
to serial processing for the rejection sampling.
This enables offloading all hashing operations in mlkem-native to a single
accelerator.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>