FIPS202: Add FIPS202 state machine to custom-FIPS202 example
We want mlkem-native to adhere to an implicit state machine for
FIPS202 whereby an initialized FIPS202 context is always in one
of the following states:

- Absorbing
- Squeezing
- Reset

A call to `init()` sets the state to `Absorbing` (from any input
state), a call to `absorb` retains the `Absorbing` state, a call
to `absorb_once` moves from `Absorbing` to `Squeezing`, and calls
to `squeezeblocks` move from `Squeezing` to itself. Finally, `reset`
moves from `Squeezing` to `Reset` (note that implementations will
likely allow `reset()` from any initialized state, but mlkem-native
does not make use of such paths).

We previously did not check that mlkem-native adhered to this discipline.

This commit extends the custom-FIPS202 example examples/custom_fips202,
wrapping the tiny-SHA3 context in a struct additionally including an
explicit state. This state is transitions as above in the glue code,
and we also add assertions checking the incoming state.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>