Avoid disabling CBMC signed-to-unsigned conversion checks
There a few places in mlkem-native where we need to convert
from signed to unsigned integers, and wish to rely on the
defined behavior that conversion into an unsigned domain with
modulus R should convert a value `x` into the unique `y` in
the range [0..R) so that `x == y (mod R)`.

However, our use of CBMC by default flags integer conversions
which change the value, and for negative inputs, signed to
unsigned conversions are such. Previously, we silenced the
corresponding proof failures by temporarily disabling the
conversion checks.

In this commit, we avoid two such instances:
- In Montgomery reduce, we want to convert an `int16_t` to
  an `uint16_t`, and the previous approach was to use the expression
  `(uint16_t)((uint32_t)a & UINT16_MAX)` which relies on the
  conversion from int16_t to uint32_t.

  Here, we instead use `a & (int32_t) UINT16_MAX`, which promotes
  `a` to `int32_t` and then does a bitwise AND with UINT16_MAX,
  thereby masking the sign bit. The result is a signed integer
  in the range of uint16_t, and hence can be cast without a
  loss of value.

- A similar approach works in mlk_ct_cmask_neg_i16.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>