Make: Detect arm-embedded architecture cross compilation target
test/mk/compiler.mk has logic for detecting the target architecture,
taking into account both the host architecture and, if present, the
cross prefix.

Previously, cross-compilation for embedded Arm targets was not
recognized, and the host architecture (e.g. AArch64 or x86_64)
chosen as the target architecture. This would lead to compilation
failures because the makefiles would later set MLK_FORCE_XXX based
on the arch, contradicting the actual target of the (cross-)compilation.

This commit fixes the issue by detecting arm embedded cross
compilation targets.

The commit also modifies the Makefiles to not fall back to the
host architecture as the target architecture if a cross prefix
is set but not recognized.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>