Set `-monitor none` for qemu-system-arm to enable abort via CTRL-C
Previously, you couldn't abort (the emulation step in)

```
  EXTRA_MAKEFILE=test/baremetal/platform/m55-an547/platform.mk tests func
```

via CTRL-C/SIGINT. It seems that this is related to qemu-system-arm
using CTRL-C/SIGINT for something related to the monitor console.

This commit adds `-monitor none` to the `qemu-system-arm` invocation
to disable the monitor and free up CTRL-C for interruption of the
emulation.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>