Allow additional CFLAGS for extra sources
Since the addition of -Wconversion and -Wsign-conversion to the default
CFLAGS, compilation of the CMSIS files for AN547 fails.

This commit allows custom Makefiles to not only provide EXTRA_SOURCES
to specify additional sources, but also EXTRA_SOURCES_CFLAGS for additional
CFLAGS, which in particular may overwrite default CFLAGS (they apply later).

This is leveraged in the case of m55-an547/platform.mk to disable
conversion and sign-conversion warnings for all CMSIS files.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>