CI: Add OpenTitan integration test
This commit adds an OpenTiten integration test to our CI which clones
OpenTitan, updates the mlkem-native dependency to the current commit, and runs
the ML-KEM tests using verilator.

Resolves https://github.com/pq-code-package/mlkem-native/issues/1242

OpenTitan requires an old version of verilator (4.210) which is compiled from
source taking around 5-10 minutes. We cache it (80 MiB) in the Github cache
to speed-up CI.

Another expensive step is the initial compilation of the hardware simulation.
Verilator commonly takes 10-15 minutes. Usually this is cached in the bazel
cache such that subsequent tests are much faster.
I considered caching this which would speed-up CI by around 10-15 minutes,
but require 3 GiB of caching. We are already tight in Github cache, so at this
point, I don't think we can afford it.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>